fix int enforcement (#443)

+++ config_explorer/src/config_explorer/capacity_planner.py
@@ -68,7 +68,7 @@ class KVCacheDetail:
         self.num_key_value_heads = model_config.num_key_value_heads
         self.head_dimension = getattr(model_config,"head_dim", None)
         if self.head_dimension is None:
-            self.head_dimension = self.hidden_size / self.num_attention_heads
+            self.head_dimension = int(self.hidden_size / self.num_attention_heads)
         # Determine attention type
         if use_mla(self.model_architecture):
             self.attention_type = AttentionType.MLA
